% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_mahalanobis.R
\name{plot_outliers_mahalanobis}
\alias{plot_outliers_mahalanobis}
\title{mahalanobis function to detect outliers}
\usage{
plot_outliers_mahalanobis(data, h = 0.5, alpha = 0.01, na.rm = TRUE,
  verbose = TRUE)
}
\arguments{
\item{data}{matrix of bivariate values from which we want to compute outliers}

\item{h}{when using the MMCD method, proportion of dataset to used in order to compute sample means and covariances}

\item{alpha}{alpha value (by default .01)}

\item{na.rm}{set whether Missing Values should be excluded (na.rm=TRUE) or not (na.rm=FALSE) - defaults to TRUE}

\item{verbose}{logical variable indicating whether text output should be generated (verbose = TRUE) or not (verbose = FALSE) - default to TRUE}
}
\value{
Returns Method, Max distance, number of outliers, Outliers positions
}
\description{
mahalanobis function to detect outliers
}
\examples{
## Run plot_outliers_mahalanobis
Sigma=matrix(c(1,.5,.5,1),2,2)
data=mvrnorm(100,mu=rep(0,2),Sigma=Sigma)
plot_outliers_mahalanobis(data=data, h=.5,na.rm=TRUE,plot=TRUE, verbose=TRUE)
}
\keyword{mahalanobis}
\keyword{outliers}
\keyword{plot}
