% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routliers-package.R
\docType{package}
\name{Routliers-package}
\alias{Routliers}
\alias{Routliers-package}
\title{Routliers: Robust outliers detection}
\description{
Detecting outliers using robust methods:the MAD for univariate outliers; MMCD for multivariate outliers (there is also the more known but less robust mahalanobis distance method)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdelacre/Routliers}
  \item Report bugs at \url{https://github.com/mdelacre/Routliers/issues}
}

}
\keyword{internal}
