% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_MAD.R
\name{outliers_MAD}
\alias{outliers_MAD}
\title{MAD function to detect outliers}
\usage{
outliers_MAD(data, constant = 1.4826, threshold = 3, na.rm = TRUE,
  plot = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{vector of numeric values from which we want to compute outliers}

\item{constant}{scale factor}

\item{threshold}{the number of MAD considered as a threshold to consider a value an outlier}

\item{na.rm}{set whether Missing Values should be excluded (na.rm=TRUE) or not (na.rm=FALSE) - defaults to TRUE}

\item{plot}{set whether results should be plotted (plot = TRUE) or not (plot = FALSE) - defaults to TRUE}

\item{verbose}{logical variable indicating whether text output should be generated (verbose = TRUE) or not (verbose = FALSE) - default to TRUE}
}
\value{
Returns median, MAD, lower MAD limit, upper MAD limit, extremely small values, extremely high values
}
\description{
MAD function to detect outliers
}
\examples{
Run outliers_MAD
outliers_MAD(data=rnorm(150), constant=1.4826,threshold=3,na.rm=TRUE,
plot=TRUE, verbose=TRUE)
}
\keyword{MAD}
\keyword{outliers}
