% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_mahalanobis.R
\name{outliers_mahalanobis}
\alias{outliers_mahalanobis}
\title{MMCD function to detect outliers}
\usage{
outliers_mahalanobis(data, h = 0.5, alpha = 0.01, na.rm = TRUE,
  plot = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{matrix of bivariate values from which we want to compute outliers}

\item{h}{when using the MMCD method, proportion of dataset to used in order to compute sample means and covariances}

\item{alpha}{alpha value (by default .01)}

\item{na.rm}{set whether Missing Values should be excluded (na.rm=TRUE) or not (na.rm=FALSE) - defaults to TRUE}

\item{plot}{set whether results should be plotted (plot = TRUE) or not (plot = FALSE) - defaults to TRUE}

\item{verbose}{logical variable indicating whether text output should be generated (verbose = TRUE) or not (verbose = FALSE) - default to TRUE}

\item{method}{method of outliers detection (MMCD or Mahalanobis distance)}
}
\value{
Returns Method, Max distance, number of outliers, Outliers positions
}
\description{
MMCD function to detect outliers
}
\examples{
## Run outliers_bivar
Sigma=matrix(c(1,.5,.5,1),2,2)
data=mvrnorm(100,mu=rep(0,2),Sigma=Sigma)
outliers_mahalanobis(data=data, h=.5,na.rm=TRUE,plot=TRUE, verbose=TRUE)
}
\keyword{MMCD}
\keyword{mahalanobis}
\keyword{outliers}
