% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_mad.R
\name{plot_outliers_mad}
\alias{plot_outliers_mad}
\title{MAD function to detect outliers}
\usage{
plot_outliers_mad(data, constant = 1.4826, threshold = 3,
  na.rm = TRUE, verbose = TRUE)
}
\arguments{
\item{data}{vector of numeric values from which we want to compute outliers}

\item{constant}{scale factor}

\item{threshold}{the number of MAD considered as a threshold to consider a value an outlier}

\item{na.rm}{set whether Missing Values should be excluded (na.rm=TRUE) or not (na.rm=FALSE) - defaults to TRUE}

\item{verbose}{logical variable indicating whether text output should be generated (verbose = TRUE) or not (verbose = FALSE) - default to TRUE}
}
\description{
MAD function to detect outliers
}
\examples{
Run plot_outliers_mad
plot_outliers_mad(data=rnorm(150), constant=1.4826,threshold=3,na.rm=TRUE,
plot=TRUE, verbose=TRUE)
}
\keyword{MAD}
\keyword{outliers}
\keyword{plot}
