% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_mad.R
\name{plot_outliers_mad}
\alias{plot_outliers_mad}
\title{Plotting function for the mad}
\usage{
plot_outliers_mad(res, x)
}
\arguments{
\item{res}{vector of numeric values from which we want to compute outliers}

\item{x}{data from which the outliers_mad function was performed}
}
\value{
None
}
\description{
plotting data and highlighting univariate outliers detected with the MAD function
}
\keyword{MAD}
\keyword{outliers}
\keyword{plot}
