% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_mahalanobis.R
\name{plot_outliers_mahalanobis}
\alias{plot_outliers_mahalanobis}
\title{Plotting function for the Mahalanobis distance approach}
\usage{
plot_outliers_mahalanobis(x, alpha = 0.01, na.rm = TRUE)
}
\arguments{
\item{x}{matrix of bivariate values from which we want to compute outliers}

\item{alpha}{nominal type I error probability (by default .01)}

\item{na.rm}{set whether Missing Values should be excluded (na.rm = TRUE) or not (na.rm = FALSE) - defaults to TRUE}
}
\value{
None
}
\description{
plotting data and highlighting multivariate outliers detected with the mahalanobis distance approach
}
\details{
plotting data and highlighting multivariate outliers detected with the MCD function
Additionnally, the plot return two regression lines: the first one including all data and
the second one including all observations but the detected outliers. It allows to observe how much the outliers
influence of outliers on the regression line.
}
\examples{
## Run plot_outliers_mahalanobis
#data(Attacks)
#SOC <- rowMeans(Attacks[,c("soc1r","soc2r","soc3r","soc4","soc5","soc6","soc7r","soc8","soc9","soc10r","soc11","soc12","soc13")])
#HSC <- rowMeans(Attacks[,22:46])
#plot_outliers_mahalanobis(x = cbind(SOC,HSC),na.rm = TRUE)

}
\keyword{mahalanobis}
\keyword{outliers}
\keyword{plot}
