% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_mcd.R
\name{plot_outliers_mcd}
\alias{plot_outliers_mcd}
\title{Plotting function for the MCD}
\usage{
plot_outliers_mcd(data, h = 0.5, alpha = 0.01, na.rm = TRUE,
  verbose = TRUE)
}
\arguments{
\item{data}{matrix of bivariate values from which we want to compute outliers}

\item{h}{when using the MMCD method, proportion of dataset to used in order to compute sample means and covariances}

\item{alpha}{alpha value (by default .01)}

\item{na.rm}{set whether Missing Values should be excluded (na.rm=TRUE) or not (na.rm=FALSE) - defaults to TRUE}

\item{verbose}{logical variable indicating whether text output should be generated (verbose = TRUE) or not (verbose = FALSE) - default to TRUE}
}
\value{
None
}
\description{
plotting data and highlighting multivariate outliers detected with the MCD function
}
\examples{
## Run plot_outliers_mcd
SOC=rowMeans(Attacks[,c("soc1r","soc2r","soc3r","soc4","soc5","soc6","soc7r","soc8","soc9","soc10r","soc11","soc12","soc13")])
outliers_mad(data=SOC)
HSC=rowMeans(Attacks[,22:46])
plot_outliers_mcd(data=cbind(SOC,HSC), h=.5,na.rm=TRUE,verbose=TRUE)
}
\keyword{MMCD}
\keyword{outliers#'}
\keyword{plot}
