% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routliers-package.R
\docType{package}
\name{Routliers-package}
\alias{Routliers}
\alias{Routliers-package}
\title{Routliers: Robust Outliers Detection}
\description{
Detecting outliers using robust methods, 
 	i.e. the Median Absolute Deviation (MAD) for univariate outliers and the Mahalanobis-Minimum Covariance Determinant (MMCD)
        for multivariate outliers.There is also the more known but less robust Mahalanobis 
	distance method, only for comparison purposes.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/mdelacre/Routliers/issues}
}

}
\author{
\strong{Maintainer}: Marie Delacre \email{marie.delacre@ulb.ac.be}

Authors:
\itemize{
  \item Olivier Klein \email{Klein.Olivier@ulb.ac.be}
}

}
\keyword{internal}
