% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routliers-package.R
\docType{package}
\name{Routliers-package}
\alias{Routliers}
\alias{Routliers-package}
\title{Routliers: Robust outliers detection}
\description{
Detecting outliers using robust methods, i.e. the MAD for univariate outliers and MMCD for multivariate outliers.There is also the more known but less robust mahalanobis distance method, only for comparison purposes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdelacre/Routliers}
  \item Report bugs at \url{https://github.com/mdelacre/Routliers/issues}
}

}
\author{
\strong{Maintainer}: Marie Delacre \email{marie.delacre@ulb.ac.be}

Authors:
\itemize{
  \item Olivier Klein \email{Klein.Olivier@ulb.ac.be}
}

}
\keyword{internal}
